api-inferencia/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
easyocr
ultralytics
pdf2image
